Removed Config class; improved file structure
