+++ livemark/__init__.py
@@ -1,8 +1,7 @@
-from .config import Config
 from .document import Document
 from .exception import LivemarkException
+from .markup import Markup
 from .plugin import Plugin
-from .processor import Processor
 from .program import program
 from .project import Project
 from .settings import VERSION as __version__
