--- livemark/processor.py
@@ -1,3 +1,0 @@
-class Processor:
-    def process(self, document):
-        return document
