+++ livemark/program/common.py
@@ -1,0 +1,1 @@
+# TODO: extract common elements
