--- tests/test_processor.py
@@ -1,11 +1,0 @@
-from livemark import Processor, Document
-
-
-# General
-
-
-def test_processor():
-    document = Document("index.md", target="index.html")
-    processor = Processor()
-    document = processor.process(document)
-    assert document
